Azuris/requirements.txt
Django==3.2.3
knox==0.1.11
contextvars==2.4
djangorestframework==3.12.4
